requirements.txt
click==7.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
psycopg2==2.8.5
python-dotenv==0.12.0
PyYAML==5.3.1
SQLAlchemy==1.3.15
Werkzeug==1.0.1
WTForms==2.2.1
